+++ src/feature/hook/use_data_user.jsx
@@ -1,14 +1,12 @@
 import { useState, useEffect } from "react";
 
-const useUserData = () => {
-  const [users, setUsers] = useState([]);
-  const [authenticatedUser, setAuthenticatedUser] = useState(null);
+const UseDataUser = () => {
 
+  const [users, setUsers] = useState([]);
   const urlApi = "https://646d4d739c677e232189e51a.mockapi.io/user";
 
   useEffect(() => {
     fetchData();
-
     async function fetchData() {
       try {
         const response = await fetch(urlApi);
@@ -20,52 +18,11 @@ const useUserData = () => {
     }
   }, []);
 
-  useEffect(() => {
-    // Aquí puedes realizar cualquier lógica adicional cuando el usuario autenticado cambie
-    console.log("Usuario autenticado:", authenticatedUser);
-  }, [authenticatedUser]);
-
-  const createUser = async (user) => {
-    try {
-      const name  = user.name;
-      const email = user.email;
-      const password = user.password;
-
-      const response = await fetch(urlApi, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        
-        body: JSON.stringify({name,email,password,}),
-      });
-      const newUser = await response.json();
-      setUsers((prevUsers) => [...prevUsers, newUser]);
-
-      console.log("datos enviados:", newUser);
-    } catch (error) {
-      console.error("Error creating user:", error);
-    }
-  };
-
-  const authenticateUser = (email, password) => {
-    const authenticatedUser = users.find(
-      (user) => user.email === email && user.password === password
-    );
-
-    if (authenticatedUser) {
-      setAuthenticatedUser(authenticatedUser);
-    } else {
-      console.log("Autenticación fallida");
-    }
-  };
 
   return {
-    users,
-    authenticatedUser,
-    createUser,
-    authenticateUser,
+    urlApi,
+    setUsers,
   };
 };
 
-export default useUserData;
+export default UseDataUser;
